Restore the state when OpenSLESRecorder::CreateAudioRecorder fails
OpenSLESRecorder will be inconsistent if CreateAudioRecorder fails.
For example, it is possible that recorder_object_ is set and
simple_buffer_queue_ is not set at the same time.
In that case, OpenSLESRecorder::DestroyAudioRecorder considers it audio
recorder is created, and tries to use simple_buffer_queue_ to reset the
state, which leads to segmentation fault.

Restore the state when OpenSLESRecorder::CreateAudioRecorder fails to
avoid such a complication.