Relicense the additions made by pixiv Inc.
A quick internal review revealed LICENSE, which claims WebRTC project is
awkward, as the additions are not contributed to WebRTC project yet.

It was also pointed out that lack of a counterpart of PATENTS may also
lead to uncertainity.

This change adds LICENSE.pixiv, which includes the copy of LICENSE with
the copyright holder replaced, and Apache-2.0 license, noting the additions
made by pixiv Inc. is available under either of the terms in LICENSE, and
Apache-2.0.

The terms copied from LICENSE are expected to make it easier if someone
wants to integrate the additions to the upstream. Apache-2.0 provides an
explicit patent grant.