Allow to call deliverRecordedData: on multiple threads
It is required to keep [RTCAudioDeviceModule deliverRecordedData:] calling
once an audio stream added. In such a case, you may invoke the method
with dummy samples on the background thread before real samples are ready,
and then invoke the method with real samples in threads handling audio.
This change covers the scenario.