Do not try to have a "perfect" scale factor in webrtc::VideoAdapter
Without this change, webrtc::VideoAdapter tries to have a scale factor
which can be represented in an easy fraction. That behavior requires
the cropped image to be aligned with scale fraction denominator multiplied
with the output alignment requirement, which can be much larger for
relatively large output alignment requirement.

The behavior does not seem to be required by any implementation so just
abandon it to avoid such a large alignment requirement.